% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{addChromosomeLines}
\alias{addChromosomeLines}
\title{Add chromosome boundaries and labels to a plot}
\usage{
addChromosomeLines(genome, borders = FALSE, lncol = "gray70", ...)
}
\arguments{
\item{genome}{`genome` or `population` corresponding to the
correlation matrix plotted on the active device}

\item{borders}{logical, should extra borders around each
chromosome block on the diagonal be added?}

\item{lncol}{character giving the colour of lines}

\item{...}{additional arguments to pass to rect if borders = TRUE}
}
\value{
Returns nothing, but adds lines separating chromosomes
to a plotted correlation heat map
}
\description{
Add chromosome boundaries and labels to a plot
}
\details{
Meant to be used after calling `corrImg`, this function
adds vertical and horizontal lines to a plot to visually separate
the chromosomes of a correlation matrix and labels the bordered
regions accordingly.
}
\author{
Chris Salahub
}
