% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomes.R
\name{asGenome}
\alias{asGenome}
\title{Convert a `data.frame` to a `genome`}
\usage{
asGenome(
  df,
  alleles,
  values = c(1, 0),
  missing = c(".", "-"),
  encoder = encode
)
}
\arguments{
\item{df}{`data.frame` with columns `mv`, `pv`, `chr`, and `pos`}

\item{alleles}{list with `length(alleles) == nrow(df)` that
gives the potential alleles for each row}

\item{values}{list of identical form to `alleles` giving numeric
values for the encoding of each possible allele}

\item{missing}{character vector giving string values to be
interpreted as missing}

\item{encoder}{function which accepts an annotation and its
possible values and returns an encoding}
}
\value{
A `genome` with locations given by `pos` across
chromosomes given by `chr` with encodings given by the values in
`mv` and `pv`
}
\description{
Convert a `data.frame` to a `genome`
}
\details{
Given a `data.frame` with annotations in a particular
structure, potentially with missing annotations denoted by a set
list of characters, this function generates the corresponding
`genome` by encoding the data and changing its format.
}
\author{
Chris Salahub
}
