% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{asPopulation}
\alias{asPopulation}
\title{Create a population object}
\usage{
asPopulation(genomes)
}
\arguments{
\item{genomes}{list of `genome`s with the same locations}
}
\value{
A population object.
}
\description{
Create a population object
}
\details{
Provided a list of `genome`s, this function strips
redundant information from the genomes to convert them to a
`population`, which has the same slots as `genome` except that
`encoding` is now `encodings` to reflect that it is a list of
encoding matrices and `marker` is added to remove redundant row
names.
}
\author{
Chris Salahub
}
