% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cope_jenk}
\alias{cope_jenk}
\title{(C57BL/6J x M. spretus)F1 x C57BL/6J Data}
\format{
A population of 205 genomes sequenced at 1394 markers with
roughly 9% of the data missing.
}
\source{
<https://www.informatics.jax.org/downloads/reports/MGI_Copeland-Jenkins_Panel.rpt>
}
\usage{
cope_jenk
}
\description{
These data sets are population objects providing the results
of backcross experiments carried out on different inbred strains
of mice processed from raw data. The inbred strains are
identified using their Jackson Laboratory codes in the title for
each data set. For complete context on each, please consult the
Mouse Genome Database (https://www.informatics.jax.org/), which
has references explaining the genetic markers, data sources, and
inbred strains.
}
\keyword{datasets}
