% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeCombination.R
\name{crossIndep}
\alias{crossIndep}
\title{Functions to simulate crossing over}
\usage{
crossIndep(probs, locs)
}
\arguments{
\item{probs}{vector of probabilities of cross overs between
markers}

\item{locs}{vector of locations of markers on a chromosome}
}
\value{
A vecotr giving the indices of cross over events.
}
\description{
Functions to simulate crossing over
}
\details{
These functions are core to the action of `meiose` by
accepting a numeric vector of probabilities of recombination and
a numeric vector of locations and returning the indices in the
location vector where cross over events occur. By default, only
a simple independent cross over function is included, but a
chi-squared based function is shown in the `customFuns` demo.
}
\author{
Chris Salahub
}
