% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomes.R
\name{encode}
\alias{encode}
\title{Simple encoding function}
\usage{
encode(annotation, alleles, values)
}
\arguments{
\item{annotation}{character vector giving observed annotations}

\item{alleles}{character vector of possible annotations}

\item{values}{numeric vector of the same length as alleles giving
the value of each allele when encoded}
}
\value{
A numeric vector of encodings.
}
\description{
Simple encoding function
}
\details{
This function takes observed annotations, a vector of
possible annotations, and a vector of values for each, and returns
an encoding given by replacing annotations with values.
}
\author{
Chris Salahub
}
