% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eucib_bss}
\alias{eucib_bss}
\title{(C57BL/6J x SPR or SEG/Pas)F1 x SPR or SEG/Pas Data}
\format{
A population of 501 genomes sequenced at 3332 markers with
roughly 46% of the data missing.
}
\source{
<https://www.informatics.jax.org/downloads/reports/MGI_EUCIB_BSS_Panel.rpt>
}
\usage{
eucib_bss
}
\description{
These data sets are population objects providing the results
of backcross experiments carried out on different inbred strains
of mice processed from raw data. The inbred strains are
identified using their Jackson Laboratory codes in the title for
each data set. For complete context on each, please consult the
Mouse Genome Database (https://www.informatics.jax.org/), which
has references explaining the genetic markers, data sources, and
inbred strains.
}
\keyword{datasets}
