% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jax_bsb}
\alias{jax_bsb}
\title{(C57BL/6J x M. spretus)F1 x C57BL/6J Data}
\format{
A population of 94 genomes sequenced at 1599 markers with no
missing data.
}
\source{
<https://www.informatics.jax.org/downloads/reports/MGI_JAX_BSB_Panel.rpt>
}
\usage{
jax_bsb
}
\description{
These data sets are population objects providing the results
of backcross experiments carried out on different inbred strains
of mice processed from raw data. The inbred strains are
identified using their Jackson Laboratory codes in the title for
each data set. For complete context on each, please consult the
Mouse Genome Database (https://www.informatics.jax.org/), which
has references explaining the genetic markers, data sources, and
inbred strains.
}
\keyword{datasets}
