% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jax_bss}
\alias{jax_bss}
\title{(C57BL/6JEi x SPRET/Ei)F1 x SPRET/EiJ Data}
\format{
A population of 94 genomes sequenced at 3639 markers with roughly
1% of the data missing.
}
\source{
<https://www.informatics.jax.org/downloads/reports/MGI_JAX_BSS_Panel.rpt>
}
\usage{
jax_bss
}
\description{
These data sets are population objects providing the results
of backcross experiments carried out on different inbred strains
of mice processed from raw data. The inbred strains are
identified using their Jackson Laboratory codes in the title for
each data set. For complete context on each, please consult the
Mouse Genome Database (https://www.informatics.jax.org/), which
has references explaining the genetic markers, data sources, and
inbred strains.
}
\keyword{datasets}
