% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeCombination.R
\name{mapHaldane}
\alias{mapHaldane}
\alias{invHaldane}
\alias{mapKosambi}
\alias{invKosambi}
\title{Genetic mapping functions}
\usage{
mapHaldane(d)

invHaldane(pr)

mapKosambi(d)

invKosambi(pr)
}
\arguments{
\item{d}{numeric vector of additive map distances}

\item{pr}{numeric vector of probabilities}
}
\value{
Returns a numeric vector of recombination probabilities
based on the corresponding map distance units.
}
\description{
These functions support the conversion of vectors
between map distance and probability of recombination.
}
\details{
Maps

These functions convert map distances (assumed to be in
centiMorgans) to probabilities of recombination or the inverse.
Only the Haldane and Kosambi distances are defined by default.
}
\section{Functions}{
\itemize{
\item \code{mapHaldane()}: Haldane's map distance

\item \code{invHaldane()}: The inverse of Haldane's map distance

\item \code{mapKosambi()}: Kosambi's map distance

\item \code{invKosambi()}: The inverse of Kosambi's map distance

}}
\author{
Chris Salahub
}
