% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeCombination.R
\name{meiose}
\alias{meiose}
\title{Simulating meiosis}
\usage{
meiose(genome, probs = NULL, crossFun = crossIndep, map = mapHaldane)
}
\arguments{
\item{genome}{a `genome` to be recombined}

\item{probs}{numeric vector of probabilities of recombination
which override distance calculations}

\item{crossFun}{function which accepts a vector of probabilities
and a vector of marker locations and returns the indices where
crossovers occur}

\item{map}{function which accepts distances between locations in
`genome` and returns the probabilities of recombination for each
distance}
}
\value{
A list of encoding sub-matrices split by chromosome with
certain rows of the encoding recombined as dictated by `probs`,
`crossFun`, and `map`.
}
\description{
Simulating meiosis
}
\details{
A helper function for `sex`, `meiose` dictates the
recombination that occurs within a `genome` and is meant to
mimic the action of meiosis in the creation of gametes. By
changing `crossFun` to reflect a given model of crossing over,
effectively any internal genetic recombination model can be
mimicked by this function.
}
\author{
Chris Salahub
}
