% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomes.R
\name{print.genome}
\alias{print.genome}
\alias{plot.genome}
\title{S3 methods for `genome`}
\usage{
\method{print}{genome}(genome)

\method{plot}{genome}(
  genome,
  chrLens,
  epch = c(1, 2, 6, 0),
  elevs = c("0 0", "0 1", "1 0", "1 1"),
  scale = FALSE,
  add.legend = FALSE,
  ...
)
}
\arguments{
\item{genome}{instance of the genome class}

\item{chrLens}{numeric vector of lengths of each chromosome, if
missing taken as the largest location on each chromosome}

\item{epch}{vector of R pch values for the potential encoding
combinations}

\item{scale}{logical: if chrLens is provided, should the locations
be scaled to [0,1] when plotted?}

\item{add.legend}{if TRUE, draw a legend in the top right corner,
if FALSE suppress legend}

\item{elev}{vector of possible encodings generated by pasting
the two columns of the encoding matrix together, defaults to
"0 0", "0 1", "1 0", "1 1"}
}
\value{
Nothing is returned, but a quick summary of the genome is
printed to console or a visualization of the genome is plotted
on the active device.
}
\description{
The `print` and `plot` methods outlined here support
the quick description and exploration of a `genome`.
}
\details{
Methods

These methods for printing and plotting genome objects
are helpful to get a quick overview of the encoding contained in
a genome and the structure of the locations encoded.
}
\section{Functions}{
\itemize{
\item \code{print(genome)}: Print method for `genome`

\item \code{plot(genome)}: Plot method for `genome`

}}
\author{
Chris Salahub
}
