% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgiMarkers}
\alias{mgiMarkers}
\title{Marker Positions in the Mouse Genome Database}
\format{
A list of 20 data frames each containing columns `loc`, `sym`,
`beg`, and `end` which report the cM location, marker symbol,
beginning coordinate in base pairs, and end coordinate in base
pairs of markers in the Mouse Genome Database.
}
\source{
<https://www.informatics.jax.org/downloads/reports/MRK_List2.rpt>
}
\usage{
mgiMarkers
}
\description{
This list of data frames provides the chromosome, marker ID, cM
positions, and base pair coordinates of genetic markers included
in the Mouse Genome Database (https://www.informatics.jax.org/).
}
\keyword{datasets}
