% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{print.population}
\alias{print.population}
\title{S3 methods for `population`}
\usage{
\method{print}{population}(population)
}
\arguments{
\item{population}{`population` object}
}
\value{
Returns nothing, but prints a quick summary of the
populations to console
}
\description{
Print method
}
\author{
Chris Salahub
}
