% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{scoreAdditive}
\alias{scoreAdditive}
\alias{scoreDominance}
\alias{scoreHomozygous}
\title{Summarizing an encoding ("scoring" the genome)}
\usage{
scoreAdditive(encoding)

scoreDominance(encoding)

scoreHomozygous(encoding)
}
\arguments{
\item{encoding}{matrix with two columns giving marker encodings}
}
\value{
A numeric vector of scores with one numeric for each row
of `encoding`.
}
\description{
These functions convert two-column matrices of
encodings into summary vectors based on an operation applied to
each row.
}
\details{
Scoring

These functions accept encoding matrices and return
numeric vectors that summarize the marker measurements at each
row.
}
\section{Functions}{
\itemize{
\item \code{scoreAdditive()}: The additive score

\item \code{scoreDominance()}: The dominance score

\item \code{scoreHomozygous()}: The homozygous score

}}
\author{
Chris Salahub
}
