% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomes.R
\name{subsetGenome}
\alias{subsetGenome}
\title{Subsetting the markers of a genome object}
\usage{
subsetGenome(genome, inds)
}
\arguments{
\item{genome}{object to be subset}

\item{inds}{indices of markers to be kept}
}
\value{
A genome object containing only the markers at `inds`,
with any chromosomes fully removed dropped as levels of
`genome$chromosome`.
}
\description{
Subsetting the markers of a genome object
}
\details{
A helper function that subsets every slot of a `genome`
and drops unused `chromosome` levels by a set of marker indices.
}
\author{
Chris Salahub
}
